Refactor server API import routes
