+++ crates/rulemorph_server/tests/cloud_scenarios.rs
@@ -1960,6 +1960,104 @@ async fn internal_import_zip_route_is_removed() -> Result<()> {
 }
 
 #[tokio::test]
+async fn internal_import_path_requires_auth_before_bundle_validation() -> Result<()> {
+    let temp = tempdir().expect("tempdir");
+    let data_dir = temp.path().join("data");
+    let store = TraceStore::new(data_dir.clone()).await?;
+    let (trace_events, _) = broadcast::channel(16);
+    let auth_dir = data_dir.join("auth");
+    fs::create_dir_all(&auth_dir).expect("create auth dir");
+    let resources = TenantResources {
+        tenant_id: "default".to_string(),
+        data_dir: data_dir.clone(),
+        rules_dir: data_dir.join("api_rules"),
+        auth_dir,
+        store: Arc::new(store),
+        api_engine: None,
+        trace_events,
+    };
+    let state = AppState {
+        default_resources: Arc::new(resources),
+        tenant_registry: None,
+        ui_source: None,
+        api_mode: ApiMode::UiOnly,
+        tenant_resolver: None,
+        internal_api_key: Some("internal-key".to_string()),
+        allow_unauth_internal: false,
+        rate_limiter: None,
+    };
+    let app = build_router(state, true);
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/internal/import")
+                .header("content-type", "application/json")
+                .body(Body::from(
+                    json!({ "bundle_path": "/definitely/not/a/rulemorph/bundle" }).to_string(),
+                ))
+                .unwrap(),
+        )
+        .await
+        .expect("response");
+    assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn internal_import_path_rejects_non_temp_bundle_path_after_auth() -> Result<()> {
+    let temp = tempdir().expect("tempdir");
+    let data_dir = temp.path().join("data");
+    let store = TraceStore::new(data_dir.clone()).await?;
+    let (trace_events, _) = broadcast::channel(16);
+    let auth_dir = data_dir.join("auth");
+    fs::create_dir_all(&auth_dir).expect("create auth dir");
+    let resources = TenantResources {
+        tenant_id: "default".to_string(),
+        data_dir: data_dir.clone(),
+        rules_dir: data_dir.join("api_rules"),
+        auth_dir,
+        store: Arc::new(store),
+        api_engine: None,
+        trace_events,
+    };
+    let state = AppState {
+        default_resources: Arc::new(resources),
+        tenant_registry: None,
+        ui_source: None,
+        api_mode: ApiMode::UiOnly,
+        tenant_resolver: None,
+        internal_api_key: Some("internal-key".to_string()),
+        allow_unauth_internal: false,
+        rate_limiter: None,
+    };
+    let app = build_router(state, true);
+    let non_temp_dir = std::env::current_dir().expect("current dir");
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/internal/import")
+                .header("authorization", "Bearer internal-key")
+                .header("content-type", "application/json")
+                .body(Body::from(
+                    json!({ "bundle_path": non_temp_dir }).to_string(),
+                ))
+                .unwrap(),
+        )
+        .await
+        .expect("response");
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+
+    Ok(())
+}
+
+#[tokio::test]
 async fn api_import_requires_internal_key_without_ui() -> Result<()> {
     let temp = tempdir().expect("tempdir");
     let data_dir = temp.path().join("data");
